done bs lesa el age

+++ client/src/pages/ItineraryList.js
@@ -4,18 +4,27 @@ import axios from "axios";
 import ItineraryItem from "../components/ItineraryItem"; // Import the ItineraryItem component
 import { useNavigate, Link } from "react-router-dom";
 import "../css/ItineraryList.css"; // Import the CSS file
+import { useCookies } from "react-cookie";
 
 const ItineraryList = () => {
   const [itineraries, setItineraries] = useState([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
   const navigate = useNavigate(); // Use useNavigate for navigation
-
+  const [cookies] = useCookies(["username", "userType"]);
+  const username = cookies.username;
+  const userType = cookies.userType; // Access the userType
   useEffect(() => {
     const fetchItineraries = async () => {
       try {
         const response = await axios.get("http://localhost:3000/itinerary"); // Replace with your API endpoint
-        setItineraries(response.data);
+
+        // Filter itineraries where the creator matches the username in the cookies
+        const filteredItineraries = response.data.filter(
+          (itinerary) => itinerary.creator === username
+        );
+
+        setItineraries(filteredItineraries); // Store the filtered itineraries in state
       } catch (err) {
         setError(err.message);
       } finally {
@@ -23,13 +32,25 @@ const ItineraryList = () => {
       }
     };
 
-    fetchItineraries();
-  }, []);
+    if (username) {
+      // Fetch itineraries only if username is available
+      fetchItineraries();
+    }
+  }, [username]);
 
   const handleDelete = async (id) => {
     try {
-      await axios.delete(`http://localhost:3000/itinerary/${id}`); // Make sure to update the endpoint
-      setItineraries(itineraries.filter((itinerary) => itinerary._id !== id)); // Update the state
+      const itinerary = await axios.get(
+        `http://localhost:3000/itinerary/${id}`
+      );
+      if (itinerary.data.bookingCounter == 0) {
+        await axios.delete(`http://localhost:3000/itinerary/${id}`); // Make sure to update the endpoint
+        setItineraries(itineraries.filter((itinerary) => itinerary._id !== id)); // Update the state
+      } else {
+        console.log(
+          `Cannot delete an itinerary with bookings ${itinerary.data.bookingCounter}`
+        );
+      }
     } catch (err) {
       setError(err.message);
     }
@@ -56,7 +77,7 @@ const ItineraryList = () => {
       <h1>Itineraries</h1>
       <div className="button-container">
         <button className="create-button" onClick={() => handleCreate()}>
-          Create New Activity
+          Create New Itinerary
         </button>
         <Link to="/home">
           <button>Back</button>
@@ -69,6 +90,7 @@ const ItineraryList = () => {
             itinerary={itinerary}
             onDelete={handleDelete}
             onEdit={handleEdit}
+            userType={userType}
           />
         ))}
       </div>
